Refactor init to await authentication
